modegen: Round display clocks up instead of down
Fixes oboe getting 29fps for vrefresh or whatever

Stops using floats here too, floats bad.

+++ src/modegen.hpp
@@ -9,4 +9,4 @@
 void generate_cvt_mode(drmModeModeInfo *mode, int hdisplay, int vdisplay,
 	float vrefresh, bool reduced, bool interlaced);
 void generate_fixed_mode(drmModeModeInfo *mode, const drmModeModeInfo *base,
-	float vrefresh);
+	int vrefresh);
